Fix broken `overlay-panel` demo in Firefox
The large image used for testing the scroll wasn't being constrained by the `max-width: 100%` rule, pushing the container out of shape and making the grid cells wrap.

See: https://bugzilla.mozilla.org/show_bug.cgi?id=975632

This fixes the demo but doesn't resolve the issue. However, we shouldn't be using raw images in the site, so this isn't worth the time to explore properly.