Move form width bounding rules
From `rb-generic-form` to `rb-fieldset`.

Setting upper bounds on form widths should ideally be done on a wrapper around the entire form, but this can't be done
properly as forms have been split into two parts (the form itself and actions) and vary in use of other wrapper
components.

Moving the bounding to fieldset-level is the least intrusive, next-best option.